pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mooaid"
version = "0.1.0"
description = "MooAId - My Opinion AI Daemon: Predicts what your opinion would likely be"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MooAId Team"}
]
keywords = ["ai", "opinion", "prediction", "fastapi", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.9.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0.1",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
mooaid = "mooaid.cli.main:app"

[project.urls]
Homepage = "https://github.com/mooaid/mooaid"
Repository = "https://github.com/mooaid/mooaid"

[tool.setuptools.packages.find]
where = ["."]
include = ["mooaid*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
